Improve OpenGL error handling, fall back to 3.3 without SSAO if OpenGL 4.1 is not available.

+++ roboschool/cpp-household/glsl/fullscreen_triangle.vert.glsl
@@ -1,10 +1,11 @@
-//#line 2 "fullscreenquad.vert.glsl"
+#version 330
+//#line 3 "fullscreen_triangle.vert.glsl"
 
 out vec2 texCoord;
 
 void main()
 {
-  uint idx = gl_VertexID % 3;
+  uint idx = uint(gl_VertexID % 3);
   vec4 pos =  vec4(
       (float( idx     &1U)) * 4.0 - 1.0, // one big triangle stratches further than -1 (not to use quad)
       (float((idx>>1U)&1U)) * 4.0 - 1.0,
